frontend/admin/llm: refactor and add history/system test
